Keep track of the total balance and display it in the UI
The GODcoin library had to be updated before any strings could be sent
across the IPC boundary. There was a bug reading UTF-8 strings that was
fixed. The balance is sent to the renderer through the IPC post_init call by
serializing the amount as a string and then creating a new Asset instance
from the string.

The synchronizer retrieves the full balance amount through an RPC call
to the blockchain and then updates the KV table in the database. The
applyBlock function will return a bool whether or not the balance should
be updated in order to minimize RPC calls every time a block is
received during real time synchronization. During startup sync,
the RPC call will be invoked once upon completion.