Display the sync status in the sidebar
By default the default sync status in the Vuex store is in progress. The
main process will update the status to complete automatically once
startup sync completes.

There's an incorrect warning from a re-export of an interface saying it
doesn't exist. I worked around it by importing the interface directly
and removing the re-export.