Store the script in the settings instead of the Vuex store
There is no need for the Vuex store to have access to the public key.
Instead, the script is sent over the IPC boundary in the post_init
method. The script can get converted to a script hash as necessary.

This information is necessary to convert the transaction data to
something displayable to the user by testing whether the transaction
was sent or received. Later, this information will be used in the
"Receive" button dialog.