Broadcast transaction after clicking the "Send" button
IPC has a transfer_funds method that is called from the renderer. The tx
is created, signed, and broadcasted from the main process which has
access to the currently connected client.

The ChainSynchronizer now emits a sync_update event which is used by
the IPC manager when a transfer_funds request is sent. This allows the
method to use this information to determine if the transaction has been
included in a block.

Broadcast errors now display a message specifying more details on what
happened.