Implement the backup page
An IPC call has been added to retrieve the public key so that the
private key WIF can be displayed. A "reveal private key" button is also
added for after they read the informational text and to prevent
accidentally revealing it if they randomly navigate to the backup page.

The GODcoin library has been updated to support KeyPair.fromSeed(seed)
for IPC serialization and deserialization.