Load the dashboard window if the settings are loaded
Rather than trying to reinitialize the entire app. Load the dashboard
window if the settings are already loaded. This is safe because the
decryption key is already in memory and the user likely wants to keep the
background synchronizer running or have a way to quickly reopen the
wallet without typing the password again