Define NODE_URL global for the client URL
The NODE_URL is defined through the GODCOIN_NODE_URL process env
variable. A globals typings file was added to let TS know what was added
globally to prevent compilation errors.

A typescript lint rule had to be turned off for JS files since it was
expected a return type which is only applicable for TS.