Update e2e tests
Added cross-env dependency to set IS_SPECTRON variable that the webpack
config will be reading. When the IS_SPECTRON variable is set, we can
provide defaults for required environment variables to ensure the build
succeeds.

Updated vue-cli-plugin-electron-builder to 2.0.0-alpha since it updates
spectron to 9.0.0 (Electron 7) and vue cli service to 4.x.x. This will
be useful when we update to Electron 7.